requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.2.3
django-crispy-forms==2.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==21.2.0
kombu==5.3.1
packaging==23.1
pika==1.3.2
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
